Re-Add `uuid` property to `TextureGroupOptions`
I thought I was wrong to include the `uuid` in `TextureGroupOptions` because the `TextureGroup` constructor doesn't use the `uuid` in the provided options. However, the `getSaveCopy` and `OutlinerElement.fromSave` functions do take `uuid` into account. So it's still important to have.

I've used some type trickery to adjust the return types of all the functions that deal with `TextureGroupOptions`-like data.