Add `node-modules-vscode-problems-patch` tool
The VSCode problems patch tool is a little tool I made that adds a `// @ts-no-check` comment to the beginning of every `.d.ts` and  `.ts` file in `node_modules`, preventing any TypeScript tools / VSCode Extensions from type-checking your node modules.

I've added it to this repository to fix the errors that typedoc is finding while checking jQuery.

It's saved as a dev dependency, so it won't get installed when users add the `blockbench-types` repo as a dependency.

I've also updated the `tsconfig.json`'s exclude to use a more universal exclude pattern for node module folders.